Rename an io.h header file to fileio.h; this may be confusing Windows.
