+++ util/LLgen/src/LLgen.g
@@ -21,7 +21,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include "types.h"
-#include "io.h"
+#include "fileio.h"
 #include "extern.h"
 #include "assert.h"
 #include "cclass.h"
