+++ util/LLgen/src/gencode.c
@@ -22,7 +22,7 @@
 #include <stdio.h>
 #include <string.h>
 # include "types.h"
-# include "io.h"
+# include "fileio.h"
 # include "extern.h"
 # include "sets.h"
 # include "assert.h"
