+++ util/LLgen/src/name.c
@@ -22,7 +22,7 @@
 # include "types.h"
 # include "extern.h"
 # include "assert.h"
-# include "io.h"
+# include "fileio.h"
 
 # ifndef NORCSID
 static string rcsid7 = "$Id$";
