+++ util/LLgen/src/reach.c
@@ -20,7 +20,7 @@
 #include <stdio.h>
 # include "types.h"
 # include "extern.h"
-# include "io.h"
+# include "fileio.h"
 # include "assert.h"
 
 # ifndef NORCSID
