+++ util/LLgen/src/tokens.g
@@ -20,7 +20,7 @@
 {
 #include <string.h>
 #include "types.h"
-# include "io.h"
+# include "fileio.h"
 # include "extern.h"
 # include "assert.h"
 # include "cclass.h"
